requirements/base.txt
Django==4.2
django-jazzmin==2.6.0
django-filter==23.1
drf-yasg==1.21.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.9
Pillow==9.4.0
python-environ==0.4.54
environs==9.5.0
psycopg2-binary==2.9.6
pre-commit==3.2.2
django-cors-headers==3.14.0
django-simple-captcha==0.5.17
django-phonenumber-field==7.1.0
phonenumberslite==8.13.11
django-recaptcha==3.0.0
django-modeltranslation==0.18.9
social-auth-app-django==5.2.0
dj-rest-auth==3.0.0
django-allauth==0.54.0

